third_party/blink/web_tests/platform/generic/fast/dom/HTMLTemplateElement/ownerDocumentXHTML-expected.txt

Test HTMLTemplateElement content ownerDocument in XHTML Document

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document instanceof HTMLDocument is false
PASS document is not templateContentOwnerDocument
PASS templateContentOwnerDocument instanceof HTMLDocument is false
PASS successfullyParsed is true

TEST COMPLETE
